+++ src/linux.rs
@@ -1,20 +1,17 @@
 use crate::{IfEvent, IpNet, Ipv4Net, Ipv6Net};
 use fnv::FnvHashSet;
 use futures::ready;
-use futures::stream::{Stream, TryStreamExt};
+use futures::stream::{FusedStream, Stream, TryStreamExt};
 use futures::StreamExt;
 use rtnetlink::constants::{RTMGRP_IPV4_IFADDR, RTMGRP_IPV6_IFADDR};
 use rtnetlink::packet::address::nlas::Nla;
 use rtnetlink::packet::{AddressMessage, RtnlMessage};
 use rtnetlink::proto::{Connection, NetlinkPayload};
-use rtnetlink::sys::{AsyncSocket, SocketAddr};
-// If "only" tokio or smol+tokio are enabled, we use tokio,
-// as smol is activated via default features.
+#[cfg(feature = "smol_socket")]
+use rtnetlink::sys::SmolSocket;
 #[cfg(feature = "tokio_socket")]
-use rtnetlink::sys::TokioSocket as Socket;
-// Otherwise, when only smol enabled, we use smol.
-#[cfg(all(not(feature = "tokio_socket"), feature = "smol_socket"))]
-use rtnetlink::sys::SmolSocket as Socket;
+use rtnetlink::sys::TokioSocket;
+use rtnetlink::sys::{AsyncSocket, SocketAddr};
 use std::collections::VecDeque;
 use std::future::Future;
 use std::io::{Error, ErrorKind, Result};
@@ -22,14 +19,22 @@ use std::net::{Ipv4Addr, Ipv6Addr};
 use std::pin::Pin;
 use std::task::{Context, Poll};
 
-pub struct IfWatcher {
-    conn: Connection<RtnlMessage, Socket>,
+#[cfg(feature = "tokio_socket")]
+/// Watches for interface changes.
+pub type TokioIfWatcher = IfWatcher<TokioSocket>;
+
+#[cfg(feature = "smol_socket")]
+/// Watches for interface changes.
+pub type SmolIfWatcher = IfWatcher<SmolSocket>;
+
+pub struct IfWatcher<T> {
+    conn: Connection<RtnlMessage, T>,
     messages: Pin<Box<dyn Stream<Item = Result<RtnlMessage>> + Send>>,
     addrs: FnvHashSet<IpNet>,
     queue: VecDeque<IfEvent>,
 }
 
-impl std::fmt::Debug for IfWatcher {
+impl<T> std::fmt::Debug for IfWatcher<T> {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         f.debug_struct("IfWatcher")
             .field("addrs", &self.addrs)
@@ -37,9 +42,13 @@ impl std::fmt::Debug for IfWatcher {
     }
 }
 
-impl IfWatcher {
+impl<T> IfWatcher<T>
+where
+    T: AsyncSocket + Unpin,
+{
+    /// Create a watcher.
     pub fn new() -> Result<Self> {
-        let (mut conn, handle, messages) = rtnetlink::new_connection_with_socket::<Socket>()?;
+        let (mut conn, handle, messages) = rtnetlink::new_connection_with_socket::<T>()?;
         let groups = RTMGRP_IPV4_IFADDR | RTMGRP_IPV6_IFADDR;
         let addr = SocketAddr::new(0, groups);
         conn.socket_mut().socket_mut().bind(&addr)?;
@@ -67,6 +76,7 @@ impl IfWatcher {
         })
     }
 
+    /// Iterate over current networks.
     pub fn iter(&self) -> impl Iterator<Item = &IpNet> {
         self.addrs.iter()
     }
@@ -87,6 +97,7 @@ impl IfWatcher {
         }
     }
 
+    /// Poll for an address change event.
     pub fn poll_if_event(&mut self, cx: &mut Context) -> Poll<Result<IfEvent>> {
         loop {
             if let Some(event) = self.queue.pop_front() {
@@ -135,4 +146,23 @@ fn iter_nets(msg: AddressMessage) -> impl Iterator<Item = IpNet> {
             None
         }
     })
+}
+
+impl<T> Stream for IfWatcher<T>
+where
+    T: AsyncSocket + Unpin,
+{
+    type Item = Result<IfEvent>;
+    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
+        Pin::into_inner(self).poll_if_event(cx).map(Some)
+    }
+}
+
+impl<T> FusedStream for IfWatcher<T>
+where
+    T: AsyncSocket + AsyncSocket + Unpin,
+{
+    fn is_terminated(&self) -> bool {
+        false
+    }
 }