+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.0.0] [Unreleased]
 
 ### Changed
-- Add `IfWatcher::poll_next`. Implement `Stream` instead of `Future` for `IfWatcher`. See [PR 23].
+- Add `IfWatcher::poll_if_event`. Implement `Stream` instead of `Future` for `IfWatcher`.
+  See [PR 23] and [PR 25].
 - Make `IfWatcher::new` synchronous. See [PR 24].
 
 [PR 23]: https://github.com/mxinden/if-watch/pull/23
 [PR 24]: https://github.com/mxinden/if-watch/pull/24
+[PR 25]: https://github.com/mxinden/if-watch/pull/25
 
 ## [1.1.1]